scoutai-backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-multipart==0.0.18
sqlalchemy==2.0.36
opencv-python==4.10.0.84
mediapipe==0.10.21
ultralytics==8.3.57
pytesseract==0.3.13
scikit-learn==1.6.1
xgboost==2.1.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
celery[redis]==5.4.0
redis==5.2.1
fpdf2==2.8.2
pydantic==2.10.4
requests==2.32.3
cloudinary==1.41.0
python-dotenv==1.0.1
